+++ R/Swrappers.R
@@ -583,7 +583,7 @@ values <- function(x, ...) {
 #' A reference to the \emph{values} of a \code{\link{DTSg}} object can be used
 #'  to modify them in place. This includes the \emph{.dateTime} column, which
 #'  serves as the object's time index. Modifying this column can therefore
-#'  endanger its integrity. In case needs to do so ever arise,
+#'  endanger the object's integrity. In case needs to do so ever arise,
 #'  \code{\link{refresh}} should be called immediately afterwards in order to
 #'  check the object's integrity.
 #'