+++ NEWS.md
@@ -1,5 +1,7 @@
 ## DTSg v0.1.3.9000
 
+* Removed `xts` from suggested packages list (already comes along with `dygraphs`)
+
 ## DTSg v0.1.3
 
 * Slightly improved vignettes and documentation