+++ NEWS.md
@@ -1,5 +1,6 @@
 ## DTSg v0.1.3.9000
 
+* Added means to measure code coverage with the help of `covr`
 * Removed `xts` from suggested packages list (already comes along with `dygraphs`)
 
 ## DTSg v0.1.3