+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Added means to measure code coverage with the help of `covr`
 * Removed `xts` from suggested packages list (already comes along with `dygraphs`)
+* Slightly improved documentation
 * Minor internal code improvements
 
 ## DTSg v0.1.3