+++ NEWS.md
@@ -3,6 +3,7 @@
 * Added `swallow` argument to the `new` method: It allows for a more ressource efficient object creation
 * Added `drop` argument to the `values` method: It allows for a ressource efficient destruction of a `DTSg` object while preserving its *values*
 * Added `timestamps` field providing the total number of timestamps
+* `print` method now ommits empty metadata fields
 * Added means to measure code coverage with the help of `covr`
 * Removed `xts` from suggested packages list (already comes along with `dygraphs`)
 * Slightly improved documentation