+++ NEWS.md
@@ -1,4 +1,6 @@
-## DTSg v0.1.3.9000
+## DTSg v0.2.0.9000
+
+## DTSg v0.2.0
 
 * Added `swallow` argument to `new` method: allows for a more ressource efficient object creation
 * Added `drop` argument to `values` method: allows for a ressource efficient destruction of a `DTSg` object while preserving its *values*